# [Nim] adapt approaches in nim_con