--- a/nim/config.nims
+++ b/nim/config.nims
@@ -1,5 +1,4 @@
 switch("opt", "speed")
-switch("cc", "gcc")
 switch("panics", "off")
 switch("passC", "-s -flto")
 switch("passL", "-s -flto")