--- a/nim/config.nims
+++ b/nim/config.nims
@@ -1,5 +1,3 @@
-switch("opt", "speed")
-switch("panics", "off")
 switch("passC", "-s -flto")
 switch("passL", "-s -flto")
 switch("overflowChecks", "off")