--- a/run.sh
+++ b/run.sh
@@ -475,7 +475,7 @@ run_nim() {
         cd ./nim &&
         if [ -z "$appendToFile" ]; then # only build on 5k run
             nimble -y install -d &&
-                nimble --verbose build -d:release
+                nimble --verbose build -d:release --cc:clang
         fi &&
         if [ $HYPER == 1 ]; then
             capture "Nim" hyperfine -r $runs -w $warmup --show-output "./related"