… nim/src/fixedtable.nim