I adapted some of what I learned when speeding up `nim_con` to the single-threaded implementation. There's a modest but consistent speedup.

I ended up with a bit of an odd duck here re: `Post = ref object` and `related: array[5, ptr Post]`. Since `Post` is a ref type, it's unintuitive that `ptr Post` would achieve another small but consistent perf boost, but it does.

`clang` consistently outperforms `gcc` with these changes, so it's back to `clang`.

`threads:off` also gives a nice boost with the other changes in this PR, whereas when I tried that previously it tanked the performance. NOTE: it was already a single-threaded implementation with the default `threads:on`, but explicitly setting `threads:off` changes some aspects of how the GC works re: its use of globals vs. thread-local storage... something like that, my technical understanding there isn't 100%.