@michaelsbradleyjr 2023-10-30
Numbers in my Azure VM/Docker look promising!

```
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 22.68ms
	Processing time (w/o IO): 23.906ms
	Processing time (w/o IO): 22.236ms
	Processing time (w/o IO): 22.364ms
	Processing time (w/o IO): 22.162ms
	Processing time (w/o IO): 22.15ms
	Processing time (w/o IO): 22.311ms
	Processing time (w/o IO): 22.317ms
	Processing time (w/o IO): 22.446ms
	Processing time (w/o IO): 22.172ms
	Processing time (w/o IO): 22.213ms
	Processing time (w/o IO): 22.235ms
	Processing time (w/o IO): 22.257ms
	 Time (mean ± σ):      52.7 ms ±   0.8 ms    [User: 41.4 ms, System: 11.3 ms]
	 Range (min … max):    51.1 ms …  53.9 ms    10 runs
	
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 318.058ms
	Processing time (w/o IO): 317.609ms
	Processing time (w/o IO): 317.572ms
	 Time (mean ± σ):     442.8 ms ±   1.2 ms    [User: 407.3 ms, System: 35.4 ms]
	 Range (min … max):   441.9 ms … 443.7 ms    2 runs
	
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 2848.807ms
	Processing time (w/o IO): 2852.795ms
	Processing time (w/o IO): 2851.366ms
	 Time (mean ± σ):      3.256 s ±  0.016 s    [User: 3.174 s, System: 0.077 s]
	 Range (min … max):    3.245 s …  3.267 s    2 runs
```