@jinyus 2023-10-30
main:
``

This PR:
```
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 22.232ms
        Processing time (w/o IO): 22.04ms
        Processing time (w/o IO): 22.093ms
        Processing time (w/o IO): 21.943ms
        Processing time (w/o IO): 22.035ms
        Processing time (w/o IO): 22.036ms
        Processing time (w/o IO): 21.955ms
        Processing time (w/o IO): 21.978ms
        Processing time (w/o IO): 22.021ms
        Processing time (w/o IO): 21.895ms
        Processing time (w/o IO): 21.944ms
        Processing time (w/o IO): 21.964ms
        Processing time (w/o IO): 21.913ms
          Time (mean ± σ):      48.8 ms ±   0.3 ms    [User: 40.3 ms, System: 8.5 ms]
          Range (min … max):    48.3 ms …  49.2 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 317.746ms
        Processing time (w/o IO): 316.601ms
        Processing time (w/o IO): 315.053ms
          Time (mean ± σ):     427.9 ms ±   1.6 ms    [User: 394.6 ms, System: 33.2 ms]
          Range (min … max):   426.8 ms … 429.0 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 2835.455ms
        Processing time (w/o IO): 2837.438ms
        Processing time (w/o IO): 2842.473ms
          Time (mean ± σ):      3.212 s ±  0.006 s    [User: 3.133 s, System: 0.072 s]
          Range (min … max):    3.207 s …  3.216 s    2 runs